seperated advanced results seperatly
