+++ backend/routes/api/team.js
@@ -1,8 +1,10 @@
 const express = require('express')
+const advancedQueryResults=require('../../middlewares/advancedQueryResults')
+const Team=require('../../schemas/Team')
 
 const { createMember, getTeam } = require('../../controller/team')
 
 const router = express.Router()
-router.route('/').get(getTeam).post(createMember);
+router.route('/').get(advancedQueryResults(Team),getTeam).post(createMember);
 
 module.exports = router;
